Move functions to @prairielearn/error package
