+++ .changeset/great-pants-hug.md
@@ -1,0 +1,5 @@
+---
+"@prairielearn/error": minor
+---
+
+Introduce `formatErrorStack` and `formatErrorStackSafe` functions
